Review feedback plus some more tests that did not have native counterparts
