+++ geode-redis/src/acceptanceTest/java/org/apache/geode/redis/internal/executor/pubsub/PubSubNativeRedisAcceptanceTest.java
@@ -23,6 +23,7 @@ public class PubSubNativeRedisAcceptanceTest extends AbstractPubSubIntegrationTe
   @ClassRule
   public static NativeRedisTestRule redis = new NativeRedisTestRule();
 
+  @Override
   public int getPort() {
     return redis.getPort();
   }
