+++ geode-redis/src/integrationTest/java/org/apache/geode/redis/CommandPipeliningIntegrationTest.java
@@ -15,87 +15,15 @@
 
 package org.apache.geode.redis;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.concurrent.Callable;
-
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.ClassRule;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import redis.clients.jedis.Jedis;
-import redis.clients.jedis.Pipeline;
 
-import org.apache.geode.redis.mocks.MockSubscriber;
-import org.apache.geode.test.awaitility.GeodeAwaitility;
-import org.apache.geode.test.junit.categories.RedisTest;
-import org.apache.geode.test.junit.rules.ExecutorServiceRule;
-
-@Category({RedisTest.class})
-public class CommandPipeliningIntegrationTest {
-  static Jedis publisher;
-  static Jedis subscriber;
-  static final int REDIS_CLIENT_TIMEOUT = 100000;
+public class CommandPipeliningIntegrationTest extends AbstractCommandPipeliningIntegrationTest {
 
   @ClassRule
   public static GeodeRedisServerRule server = new GeodeRedisServerRule();
 
-  @ClassRule
-  public static ExecutorServiceRule executor = new ExecutorServiceRule();
-
-  @BeforeClass
-  public static void setUp() {
-    subscriber = new Jedis("localhost", server.getPort(), REDIS_CLIENT_TIMEOUT);
-    publisher = new Jedis("localhost", server.getPort(), REDIS_CLIENT_TIMEOUT);
-  }
-
-  @AfterClass
-  public static void tearDown() {
-    subscriber.close();
-    publisher.close();
-  }
-
+  @Override
   public int getPort() {
     return server.getPort();
-  }
-
-  @Test
-  public void whenPipelining_commandResponsesAreNotCorrupted() {
-    List<String> expectedMessages = Arrays.asList("hello");
-
-    MockSubscriber mockSubscriber = new MockSubscriber();
-
-    Runnable runnable = () -> {
-      subscriber.subscribe(mockSubscriber, "salutations");
-    };
-
-    Thread subscriberThread = new Thread(runnable);
-    subscriberThread.start();
-    waitFor(() -> mockSubscriber.getSubscribedChannels() == 1);
-
-    publisher.sadd("foo", "bar");
-
-    // Publish and smembers in a pipeline
-    Pipeline pipe = publisher.pipelined();
-
-    pipe.publish("salutations", "hello");
-    pipe.smembers("foo");
-
-    List<Object> responses = pipe.syncAndReturnAll();
-
-    mockSubscriber.unsubscribe("salutations");
-    waitFor(() -> mockSubscriber.getSubscribedChannels() == 0);
-    waitFor(() -> !subscriberThread.isAlive());
-
-    assertThat(mockSubscriber.getReceivedMessages()).isEqualTo(expectedMessages);
-  }
-
-  private void waitFor(Callable<Boolean> booleanCallable) {
-    GeodeAwaitility.await()
-        .ignoreExceptions() // ignoring socket closed exceptions
-        .until(booleanCallable);
   }
 }
