+++ geode-redis/src/integrationTest/java/org/apache/geode/redis/internal/executor/connection/AbstractPingIntegrationTest.java
@@ -30,7 +30,7 @@ import org.apache.geode.test.dunit.rules.RedisPortSupplier;
 
 public abstract class AbstractPingIntegrationTest implements RedisPortSupplier {
 
-  protected static int REDIS_CLIENT_TIMEOUT =
+  protected static final int REDIS_CLIENT_TIMEOUT =
       Math.toIntExact(GeodeAwaitility.getTimeout().toMillis());
   protected static Jedis jedis;
