+++ geode-redis/src/integrationTest/java/org/apache/geode/redis/internal/executor/key/AbstractExistsIntegrationTest.java
@@ -33,7 +33,7 @@ public abstract class AbstractExistsIntegrationTest implements RedisPortSupplier
   private Jedis jedis;
   private Jedis jedis2;
   private Jedis jedis3;
-  private static int REDIS_CLIENT_TIMEOUT =
+  private static final int REDIS_CLIENT_TIMEOUT =
       Math.toIntExact(GeodeAwaitility.getTimeout().toMillis());
 
   @Before
