+++ geode-redis/src/integrationTest/java/org/apache/geode/redis/internal/executor/key/AbstractExpireAtIntegrationTest.java
@@ -28,7 +28,8 @@ import org.apache.geode.test.dunit.rules.RedisPortSupplier;
 public abstract class AbstractExpireAtIntegrationTest implements RedisPortSupplier {
 
   private Jedis jedis;
-  private static int REDIS_CLIENT_TIMEOUT = 10000000;
+  private static final int REDIS_CLIENT_TIMEOUT =
+      Math.toIntExact(GeodeAwaitility.getTimeout().toMillis());
   private static String key = "key";
   private static String value = "value";
   private long unixTimeStampInTheFutureInSeconds;
