+++ geode-redis/src/integrationTest/java/org/apache/geode/redis/internal/executor/key/AbstractPTTLIntegrationTest.java
@@ -30,7 +30,7 @@ import org.apache.geode.test.dunit.rules.RedisPortSupplier;
 public abstract class AbstractPTTLIntegrationTest implements RedisPortSupplier {
 
   private Jedis jedis;
-  private static int REDIS_CLIENT_TIMEOUT =
+  private static final int REDIS_CLIENT_TIMEOUT =
       Math.toIntExact(GeodeAwaitility.getTimeout().toMillis());
 
   @Before
