+++ geode-redis/src/integrationTest/java/org/apache/geode/redis/internal/executor/key/AbstractPersistIntegrationTest.java
@@ -32,7 +32,7 @@ public abstract class AbstractPersistIntegrationTest implements RedisPortSupplie
 
   private Jedis jedis;
   private Jedis jedis2;
-  private static int REDIS_CLIENT_TIMEOUT =
+  private static final int REDIS_CLIENT_TIMEOUT =
       Math.toIntExact(GeodeAwaitility.getTimeout().toMillis());
 
   @Before
